Add "rb" in load_model. So reading linux generated model on windows becomes ok
Why not doing so in read_problem of svm-train.c.

Reasons of not adding "b" in svm-train.c:

From Xiang-Rui Wang:

When I use fopen() with "b" and use fgetc(), it turns out 2 chars,
'\r'(0x0d) and '\n'(0x0a)
But if I use fopen() withou 'b', only '\n' will be gotten.

And the reason libsvm can work well on Windows is because we use
fscanf() to read strings.
fscanf() will treat both '\r' and '\n' as white space and then ignore them.
there is one place we use fgetc(), but we only determine if the char
equal'\n', ':', and EOF.
and '\r'  will be treated as other char and ignored.
In other words, the newline is wrongly interpreted as two characters,
but load_model() will work well by acidentally.

Reason of adding "rb" in load_model:

From Rong-En Fan:

We use ftell+fseek in load_model and
we open model in "text" mode. According to
MSDN:

ftell returns the current file position. The value returned by ftell may
not reflect the physical byte offset for streams opened in text mode,
because text mode causes carriage return linefeed translation. Use
ftell with fseek to return to file locations correctly.

However, ftell+fseek do not work as expected.
Also see

http://bugs.python.org/issue1396471

In short, we want use binary mode in load_model,
but text mode in save_model as 1) we want
to produce pure-text model file and 2)
we want to load models that are generated
by other platform.