Earlier we used call(), which is similar to
Popen(cmd, shell = True, stdout = PIPE).wait()

According to python doc:

Popen.wait()

Warning
This will deadlock if the child process generates enough output to a stdout or stderr pipe such that it blocks waiting for the OS pipe buffer to accept more data. Use communicate() to avoid that.

Using the same data, by call() or wait(), the
program hangs if the # classes is large.

More recently, we replaced call() with

Popen(cmd, shell = True, stdout = PIPE)

The problem is that the testing data scaling 
starts before the training is finished!!
(because we don't have wait())
Then the prediction procedure starts before the
model is generated.

So we now use

Popen(cmd, shell = True, stdout = PIPE).communicate()

It waits until the procedure stops and shouldn't have the 
deadlock problem