Matlab: include 'svm.h' instead of '../svm.h'
The Matlab Makefile already has .. on the -I option for the compiler
so this was not required.  This was only required for the make.m Matlab
and Octave script, so it was modified to include .. on the search path.

This change is required to build the mex files using a libsvm library
different than the one in this distribution.  Maybe a system-wide install
of libsvm, or just another version for testing.