Fix an issue where select_working_set() does not work correctly when gradient (G) has extreme values such as INF.
In the earlier change bc92501d07bbd49737c011d9eb99c20d8fda8745 in 2005,
we remove the check of "Gmin_idx==-1" in select_working_set. 
Instead we checking only "Gmax+Gmax2 < eps". In theory this is ok, but
if for example G is INF, it may happen that Gmax = -INF and
Gmax2 = INF, and Nan < eps does not give the correct return value 1.

This issue was identified by users from github. The issue can be produced
by for example

> ./train -g 4.826430398760473E15 -n 0.25733100650382557 -c 874.7836937317443 -s 2 -t 1 -m 200000 -e 0.8230927132407901 -d 1481648982 heart_scale

or

> ./svm-train -g 1E15 -s 2 -t 1 -d 1000000000 heart_scale

This fix was prepared with the help of Po-Wei Wang