Fix libsvm name for Darwin
On macOS, shared libraries are expected to be installed with the naming
convention

    lib$NAME.$SOVERSION.dylib

We also prefix the library install name with `@rpath` to indicate to the
dynamic linker that it should use RPATH to look for `libsvm`. Without
this, the dynamic linker will only search the calling process' current
working directory, which will, in general, not contain `libsvm`.