Fix bug that could cause weights for the positive and negative classes to be incorrectly swapped.
LIBSVM supports labels like `0` and `1`. For example, in the `Solver::get_C` function, the check is `(y[i] > 0)? Cp : Cn`.

However, the `svm_group_classes` function only checks for `-1` and `1`. If instead, the labels are `0` and `1`, that could cause the classes to be in the wrong order, which would cause the class weights to be in the wrong order.